Remove the separate conjure_rule_set crate
