+++ Cargo.toml
@@ -19,7 +19,7 @@ members = [
     "crates/uniplate",
     "solvers/kissat",
     "solvers/minion",
-    "solvers/chuffed", "crates/conjure_rule_sets",
+    "solvers/chuffed",
 ]
 
 [workspace.lints.clippy]
