+++ conjure_oxide/Cargo.toml
@@ -11,7 +11,6 @@ walkdir = "2.4.0"
 [dependencies]
 conjure_core = {path = "../crates/conjure_core" }
 conjure_rules = {path = "../crates/conjure_rules" }
-conjure_rule_sets = {path = "../crates/conjure_rule_sets" }
 
 serde = { version = "1.0.196", features = ["derive"] }
 serde_json = "1.0.113"
