--- crates/conjure_rule_sets/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "conjure_rule_sets"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-conjure_core = {path= "../conjure_core"}
-conjure_rules = {path= "../conjure_rules"}
-conjure_rules_proc_macro = {path= "../conjure_rules_proc_macro"}
-linkme = "0.3.22"
-
-[lints]
-workspace = true
